backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.13.0",
    "djangorestframework>=3.16.1",
    "pillow>=12.1.1",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.2.0",
    "django-stubs>=5.2.9",
    "djangorestframework-stubs>=3.16.8",
    "mypy>=1.19.1",
    "nox-uv>=0.7.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.12.0",
    "ruff>=0.15.4",
]
prod = [
    "gunicorn>=25.1.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"

[tool.mypy]
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]
strict=true
ignore_missing_imports = true
check_untyped_defs = true

[tool.django-stubs]
django_settings_module = "core.settings"